+++ packages/components/CHANGELOG.md
@@ -18,6 +18,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
 
 - `Link Pure`: focus border of slotted anchor 
    ([#4285](https://github.com/porsche-design-system/porsche-design-system/pull/4285))
+- `Accessibility`: improved internal ARIA attribute handling for various components, e.g. `Banner`, `Inline Notification` and `Input` elements.
+  ([#4274](https://github.com/porsche-design-system/porsche-design-system/pull/4274))
 
 ## [4.0.0-rc.1] - 2026-04-07